self review clean up
