+++ OsheagaGo/OsheagaGo/Design System/Colors.swift
@@ -15,9 +15,11 @@ extension ColorPrimitive {
     static var white_universal: Self {
         make(withAssetCatalogName: "white_universal")
     }
+
     static var white_0_ink_750: Self {
         make(withAssetCatalogName: "white 0 > ink 750")
     }
+    
     static var white_100_ink_550: Self {
         make(withAssetCatalogName: "white 100 > ink 550")
     }
